Title: Renamed library to FRobs_RL
Authored-by: Jose Fajardo <jmfajardod@unal.edu.co>
