+++ src/frobs_rl/models/ddpg.py
@@ -7,9 +7,9 @@ import gym
 from gym.envs.registration import register
 import torch as th
 import stable_baselines3
-from m2rl.common import ros_params
-from m2rl.models import basic_model
-from m2rl.models.utils import get_policy_kwargs, get_action_noise
+from frobs_rl.common import ros_params
+from frobs_rl.models import basic_model
+from frobs_rl.models.utils import get_policy_kwargs, get_action_noise
 
 # ROS packages required
 import rospy
@@ -31,7 +31,7 @@ from stable_baselines3.common.callbacks import BaseCallback
 
 class DDPG(basic_model.BasicModel):
 
-    def __init__(self, env, save_model_path, log_path, config_file_pkg="m2rl", config_filename="ddpg_config.yaml", ns="/") -> None:
+    def __init__(self, env, save_model_path, log_path, config_file_pkg="frobs_rl", config_filename="ddpg_config.yaml", ns="/") -> None:
         """
         DDPG constructor.
 
@@ -39,7 +39,7 @@ class DDPG(basic_model.BasicModel):
         @param save_model_path: The path to save the model.
         @param log_path: The path to save the log.
 
-        @param config_file_pkg: The package where the config file is located. Default: m2rl.
+        @param config_file_pkg: The package where the config file is located. Default: frobs_rl.
         @param config_filename: The name of the config file. Default: ddpg_config.yaml.
         @param ns: The namespace of the ROS parameters. Default: "/".
