+++ src/frobs_rl/templates/CustomRobotEnv.py
@@ -3,14 +3,14 @@
 import gym
 from gym import spaces
 from gym.envs.registration import register
-from m2rl.common import ros_gazebo
-from m2rl.common import ros_controllers
-from m2rl.common import ros_node
-from m2rl.common import ros_launch
-from m2rl.common import ros_params
-from m2rl.common import ros_urdf
-from m2rl.common import ros_spawn
-from m2rl.envs import robot_BasicEnv
+from frobs_rl.common import ros_gazebo
+from frobs_rl.common import ros_controllers
+from frobs_rl.common import ros_node
+from frobs_rl.common import ros_launch
+from frobs_rl.common import ros_params
+from frobs_rl.common import ros_urdf
+from frobs_rl.common import ros_spawn
+from frobs_rl.envs import robot_BasicEnv
 import rospy
 
 """
@@ -19,7 +19,7 @@ robot environment.
 """
 register(
         id='CustomRobotEnv-v0',
-        entry_point='m2rl.templates.CustomRobotEnv:CustomRobotEnv',
+        entry_point='frobs_rl.templates.CustomRobotEnv:CustomRobotEnv',
         max_episode_steps=10000,
     )
