+++ src/frobs_rl/templates/CustomTaskEnv.py
@@ -3,19 +3,19 @@
 import gym
 from gym import spaces
 from gym.envs.registration import register
-from m2rl.common import ros_gazebo
-from m2rl.common import ros_controllers
-from m2rl.common import ros_node
-from m2rl.common import ros_launch
-from m2rl.common import ros_params
-from m2rl.common import ros_urdf
-from m2rl.common import ros_spawn
-from m2rl.templates import CustomRobotEnv
+from frobs_rl.common import ros_gazebo
+from frobs_rl.common import ros_controllers
+from frobs_rl.common import ros_node
+from frobs_rl.common import ros_launch
+from frobs_rl.common import ros_params
+from frobs_rl.common import ros_urdf
+from frobs_rl.common import ros_spawn
+from frobs_rl.templates import CustomRobotEnv
 import rospy
 
 register(
         id='CustomTaskEnv-v0',
-        entry_point='m2rl.templates.CustomTaskEnv:CustomTaskEnv',
+        entry_point='frobs_rl.templates.CustomTaskEnv:CustomTaskEnv',
         max_episode_steps=10000,
     )
